# src/Portwise.Web/openapi/portwise_v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Portwise API",
    "description": "Personal portfolio strategy reference API.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/stocks/{securityCode}/{exchangeCode}/model-parameters": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "Returns the active model parameters for one stock.",
        "parameters": [
          {
            "name": "securityCode",
            "in": "path",
            "description": "Six-digit A-share security code.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exchangeCode",
            "in": "path",
            "description": "Exchange code: SSE, SZSE or BSE.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockModelParameterSet"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModelParameterSet"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModelParameterSet"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/model-parameters": {
      "post": {
        "tags": [
          "Stocks"
        ],
        "summary": "Creates a new model-parameter version for one stock.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStockModelParametersRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStockModelParametersRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStockModelParametersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockModelParameterSet"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModelParameterSet"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockModelParameterSet"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{securityCode}/{exchangeCode}/price-observations/sync": {
      "post": {
        "tags": [
          "Stocks"
        ],
        "summary": "Synchronizes the latest price observation for one stock.",
        "parameters": [
          {
            "name": "securityCode",
            "in": "path",
            "description": "Six-digit A-share security code.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exchangeCode",
            "in": "path",
            "description": "Exchange code: SSE, SZSE or BSE.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockPriceObservationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockPriceObservationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockPriceObservationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{securityCode}/{exchangeCode}/dividend-events/sync": {
      "post": {
        "tags": [
          "Stocks"
        ],
        "summary": "Synchronizes dividend events for one stock.",
        "parameters": [
          {
            "name": "securityCode",
            "in": "path",
            "description": "Six-digit A-share security code.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exchangeCode",
            "in": "path",
            "description": "Exchange code: SSE, SZSE or BSE.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDividendEventResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDividendEventResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDividendEventResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{securityCode}/{exchangeCode}/financial-snapshots/sync": {
      "post": {
        "tags": [
          "Stocks"
        ],
        "summary": "Synchronizes financial snapshots for one stock.",
        "parameters": [
          {
            "name": "securityCode",
            "in": "path",
            "description": "Six-digit A-share security code.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exchangeCode",
            "in": "path",
            "description": "Exchange code: SSE, SZSE or BSE.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockFinancialSnapshotResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockFinancialSnapshotResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockFinancialSnapshotResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{securityCode}/{exchangeCode}/analysis": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "Returns the current recommendation analysis for one stock.",
        "parameters": [
          {
            "name": "securityCode",
            "in": "path",
            "description": "Six-digit A-share security code.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exchangeCode",
            "in": "path",
            "description": "Exchange code: SSE, SZSE or BSE.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StockRecommendationResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockRecommendationResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockRecommendationResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "SaveStockModelParametersRequest": {
        "required": [
          "securityCode",
          "exchangeCode",
          "modelVersion",
          "strongBuyYieldThreshold",
          "accumulationYieldThreshold",
          "partialTrimYieldThreshold",
          "aggressiveTrimYieldThreshold",
          "strongBuyBudgetRatio",
          "accumulateBudgetRatio",
          "partialTrimRatio",
          "aggressiveTrimRatio",
          "maxSecurityWeight",
          "maxSectorWeight",
          "cashReserveRatio",
          "maxSingleTradeAmount",
          "maxPeriodBudgetAmount",
          "transactionFeeRatio",
          "minimumTransactionFeeAmount",
          "tradingLotSize",
          "effectiveFromDate"
        ],
        "type": "object",
        "properties": {
          "securityCode": {
            "type": "string",
            "description": "Six-digit security code."
          },
          "exchangeCode": {
            "type": "string",
            "description": "Exchange code."
          },
          "modelVersion": {
            "type": "string",
            "description": "Client-defined model version label."
          },
          "strongBuyYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for strong-buy recommendations.",
            "format": "double"
          },
          "accumulationYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for accumulation recommendations.",
            "format": "double"
          },
          "partialTrimYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for partial-trim recommendations.",
            "format": "double"
          },
          "aggressiveTrimYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for aggressive-trim recommendations.",
            "format": "double"
          },
          "strongBuyBudgetRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Budget ratio for strong-buy recommendations.",
            "format": "double"
          },
          "accumulateBudgetRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Budget ratio for accumulation recommendations.",
            "format": "double"
          },
          "partialTrimRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Position ratio for partial trims.",
            "format": "double"
          },
          "aggressiveTrimRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Position ratio for aggressive trims.",
            "format": "double"
          },
          "maxSecurityWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum security portfolio weight.",
            "format": "double"
          },
          "maxSectorWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum sector portfolio weight.",
            "format": "double"
          },
          "cashReserveRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Portfolio cash reserve ratio.",
            "format": "double"
          },
          "maxSingleTradeAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum amount for one trade.",
            "format": "double"
          },
          "maxPeriodBudgetAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum budget for one period.",
            "format": "double"
          },
          "transactionFeeRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Transaction fee ratio.",
            "format": "double"
          },
          "minimumTransactionFeeAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Minimum transaction fee amount.",
            "format": "double"
          },
          "tradingLotSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Trading lot size in shares.",
            "format": "int32"
          },
          "effectiveFromDate": {
            "type": "string",
            "description": "Date from which the parameters apply.",
            "format": "date"
          }
        },
        "description": "Request to create a versioned model-parameter set for one stock."
      },
...
      "StockDividendEventResult": {
        "required": [
          "dividendEventId",
          "securityCode",
          "exchangeCode",
          "dividendPerShare",
          "dividendTypeCode",
          "dividendStatusCode",
          "announcementDate",
          "exDividendDate",
          "paymentDate",
          "isSpecialDividend",
          "publishedAt",
          "capturedAt",
          "dataSource",
          "sourceRecordId",
          "dataQualityCode"
        ],
        "type": "object",
        "properties": {
          "dividendEventId": {
            "type": "string",
            "description": "Identifier of the persisted dividend event.",
            "format": "uuid"
          },
          "securityCode": {
            "type": "string",
            "description": "Security code."
          },
          "exchangeCode": {
            "type": "string",
            "description": "Exchange code."
          },
          "dividendPerShare": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Dividend amount per share.",
            "format": "double"
          },
          "dividendTypeCode": {
            "type": "string",
            "description": "Normalized dividend type code."
          },
          "dividendStatusCode": {
            "type": "string",
            "description": "Normalized dividend status code."
          },
          "announcementDate": {
            "type": [
              "null",
              "string"
            ],
            "description": "Announcement date, when available.",
            "format": "date"
          },
          "exDividendDate": {
            "type": [
              "null",
              "string"
            ],
            "description": "Ex-dividend date, when available.",
            "format": "date"
          },
          "paymentDate": {
            "type": [
              "null",
              "string"
            ],
            "description": "Payment date, when available.",
            "format": "date"
          },
          "isSpecialDividend": {
            "type": "boolean",
            "description": "Whether the dividend is special."
          },
          "publishedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "Publication timestamp, when available.",
            "format": "date-time"
          },
          "capturedAt": {
            "type": "string",
            "description": "UTC capture timestamp.",
            "format": "date-time"
          },
          "dataSource": {
            "type": "string",
            "description": "Source-system name."
          },
          "sourceRecordId": {
            "type": "string",
            "description": "Source-system record identifier."
          },
          "dataQualityCode": {
            "type": "string",
            "description": "Normalized data quality code."
          }
        },
        "description": "Persisted dividend event returned by the API."
      },
...
      "StockModelParameterSet": {
        "required": [
          "modelParameterSetId",
          "securityCode",
          "exchangeCode",
          "modelVersion",
          "strongBuyYieldThreshold",
          "accumulationYieldThreshold",
          "partialTrimYieldThreshold",
          "aggressiveTrimYieldThreshold",
          "strongBuyBudgetRatio",
          "accumulateBudgetRatio",
          "partialTrimRatio",
          "aggressiveTrimRatio",
          "maxSecurityWeight",
          "maxSectorWeight",
          "cashReserveRatio",
          "maxSingleTradeAmount",
          "maxPeriodBudgetAmount",
          "transactionFeeRatio",
          "minimumTransactionFeeAmount",
          "tradingLotSize",
          "effectiveFromDate"
        ],
        "type": "object",
        "properties": {
          "modelParameterSetId": {
            "type": "string",
            "description": "Identifier of the parameter set.",
            "format": "uuid"
          },
          "securityCode": {
            "type": "string",
            "description": "Security code."
          },
          "exchangeCode": {
            "type": "string",
            "description": "Exchange code."
          },
          "modelVersion": {
            "type": "string",
            "description": "Model version label."
          },
          "strongBuyYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for strong-buy recommendations.",
            "format": "double"
          },
          "accumulationYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for accumulation recommendations.",
            "format": "double"
          },
          "partialTrimYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for partial-trim recommendations.",
            "format": "double"
          },
          "aggressiveTrimYieldThreshold": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Yield threshold for aggressive-trim recommendations.",
            "format": "double"
          },
          "strongBuyBudgetRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Budget ratio for strong-buy recommendations.",
            "format": "double"
          },
          "accumulateBudgetRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Budget ratio for accumulation recommendations.",
            "format": "double"
          },
          "partialTrimRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Position ratio for partial trims.",
            "format": "double"
          },
          "aggressiveTrimRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Position ratio for aggressive trims.",
            "format": "double"
          },
          "maxSecurityWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum security portfolio weight.",
            "format": "double"
          },
          "maxSectorWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum sector portfolio weight.",
            "format": "double"
          },
          "cashReserveRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Portfolio cash reserve ratio.",
            "format": "double"
          },
          "maxSingleTradeAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum amount for one trade.",
            "format": "double"
          },
          "maxPeriodBudgetAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Maximum budget for one period.",
            "format": "double"
          },
          "transactionFeeRatio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Transaction fee ratio.",
            "format": "double"
          },
          "minimumTransactionFeeAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Minimum transaction fee amount.",
            "format": "double"
          },
          "tradingLotSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Trading lot size in shares.",
            "format": "int32"
          },
          "effectiveFromDate": {
            "type": "string",
            "description": "Date from which the parameters apply.",
            "format": "date"
          }
        },
        "description": "Effective model parameters for one configured stock."
      },
      "StockPriceObservationResult": {
        "required": [
          "priceObservationId",
          "securityCode",
          "exchangeCode",
          "tradingDate",
          "closePrice",
          "priceObservedAt",
          "dataSource",
          "sourceRecordId",
          "dataQualityCode"
        ],
        "type": "object",
        "properties": {
          "priceObservationId": {
            "type": "string",
            "description": "Identifier of the persisted observation.",
            "format": "uuid"
          },
          "securityCode": {
            "type": "string",
            "description": "Security code."
          },
          "exchangeCode": {
            "type": "string",
            "description": "Exchange code."
          },
          "tradingDate": {
            "type": "string",
            "description": "Trading date.",
            "format": "date"
          },
          "closePrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Closing price.",
            "format": "double"
          },
          "priceObservedAt": {
            "type": "string",
            "description": "UTC observation timestamp.",
            "format": "date-time"
          },
          "dataSource": {
            "type": "string",
            "description": "Source-system name."
          },
          "sourceRecordId": {
            "type": "string",
            "description": "Source-system record identifier."
          },
          "dataQualityCode": {
            "type": "string",
            "description": "Normalized data quality code."
          }
        },
        "description": "Persisted price observation returned by the API."
      },
      "StockRecommendationResult": {
        "required": [
          "analysis",
          "suggestedBuyShares",
          "suggestedSellShares",
          "suggestedTradeAmount",
          "estimatedTransactionFeeAmount"
        ],
        "type": "object",
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/StockAnalysisResult"
          },
          "suggestedBuyShares": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Suggested buy quantity in shares.",
            "format": "int32"
          },
          "suggestedSellShares": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Suggested sell quantity in shares.",
            "format": "int32"
          },
          "suggestedTradeAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Suggested trade amount.",
            "format": "double"
          },
          "estimatedTransactionFeeAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Estimated transaction fee.",
            "format": "double"
          }
        },
        "description": "Recommendation and suggested trade quantities for one stock."
      },
...
    }
  },
...
}